fix(metrics+backend): loosen prom-client Registry types and add ambient root-bus shim; stabilize paths
