+++ packages/backend-metrics/src/helpers.ts
@@ -1,0 +1,11 @@
+export {
+  createCounter,
+  createHistogram,
+  ensureDefaultMetrics,
+  setupDefaultMetrics,
+  getRegistry,
+  getMetricsText,
+  getOpenMetricsText,
+  mergeRegistries,
+  type CollectDefaultsOptions,
+} from "./registry.js";
