+++ packages/backend-metrics/src/types.ts
@@ -1,0 +1,3 @@
+export type { CollectDefaultsOptions } from "./registry.js";
+export type { WithMetricsOptions, WithMetricsResult } from "./middleware.js";
+export type { MetricsRouterOptions } from "./router.js";
